+++ src/index.js
@@ -1,7 +1,9 @@
 import TerraformParser from './parser/TerraformParser';
 import TerraformMetadata from './metadata/TerraformMetadata';
+import TerraformRender from './render/TerraformRender';
 
-export default {
+export {
   TerraformParser,
   TerraformMetadata,
+  TerraformRender,
 };