+++ tests/tf/resourceDataExample.tf
@@ -1,4 +1,5 @@
 resource "aws_instance" "web" {
-  ami           = data.aws_ami.web.id
-  instance_type = "t1.micro"
+    ami = "data.aws_ami.web.id"
+    instance_type = "t1.micro"
 }
+